removed duplicated functions
