+++ pyclesperanto/_functionalities.py
@@ -263,10 +263,7 @@ def imshow(
         plt.title(title)
 
 
-def operations(
-    must_have_categories: Optional[list] = None,
-    must_not_have_categories: Optional[list] = None,
-) -> dict:
+def operations(must_have_categories : list = None, must_not_have_categories : list = None) -> dict:
     """Retrieve a dictionary of operations, which can be filtered by annotated categories.
 
     Parameters
@@ -288,51 +285,46 @@ def operations(
     result = {}
 
     from inspect import getmembers, isfunction
-
     import pyclesperanto as cle
 
     # retrieve all operations and cache the result for later reuse
-    operation_list = []
     if not hasattr(operations, "_all") or operations._all is None:
-        operation_list = getmembers(cle, isfunction)
+        operations._all = getmembers(cle, isfunction)
 
     # filter operations according to given constraints
-    for operation_name, operation in operation_list:
+    for operation_name, operation in operations._all:
         keep_it = True
         if hasattr(operation, "categories") and operation.categories is not None:
             if must_have_categories is not None:
-                if not all(
-                    item in operation.categories for item in must_have_categories
-                ):
+                if not all(item in operation.categories for item in must_have_categories):
                     keep_it = False
 
             if must_not_have_categories is not None:
-                if any(
-                    item in operation.categories for item in must_not_have_categories
-                ):
+                if any(item in operation.categories for item in must_not_have_categories):
                     keep_it = False
         else:
             if must_have_categories is not None:
                 keep_it = False
-        if keep_it:
+        if (keep_it):
             result[operation_name] = operation
 
     return result
 
+def operation(name : str):
+    """Returns a function from the pyclesperanto package
 
-def list_operations(search_term=None):
-    ops = operations(search_term)
-    for name in ops:
-        func = ops[name]
-        if hasattr(func, "fullargspec"):
-            print(
-                name
-                + "("
-                + str(func.fullargspec.args)
-                .replace("[", "")
-                .replace("]", "")
-                .replace("'", "")
-                + ")"
-            )
-        else:
-            print(name)
+    Parameters
+    ----------
+    name : str
+        name of the operation
+
+    Returns
+    -------
+        function
+    """
+    dict = operations()
+    return dict[name]
+
+def search_operation_names(name):
+    return [a for a in list(operations().keys()) if name in a]
+
